reader: undo setting of default reader class as ProxyReader
Specifically, when the Reader constructor call is determining what
class to construct it selects a default class of Reader or whatever
was explicitly passed into it.

If it was a "type" as a string the the look for that will override
the previous type.

When the file was being reformatted for standardization it got a
default added to the type string-based look up. This resulted in
any non string-based type being constructed as a ProxyReader type,
instead of the custom type that was selected.

As npm uses custom fstream readers, this resulted in npm and its
tests exploding.