Revert "package: Fix incorrect version specifier for mkdirp"
This reverts commit c16d373489794edff044a3977c97d883650f1db9.

The original commit changed ^0.5.0 to >=0.5 0

And it turns out that despite A) making the versions travis was configured
to work with explode and B) looking wrong, it is, in fact, correct.

What it's saying is >=0.5 AND a major version of 0. Which is what we'd
probably want. Though it's interesting to note that it's not semantically
equivalent to ^0.5.0. ¯\_(ツ)_/¯