read: account for entries changed after _read
If this.entries is changed after _read() has been called, we will be
out of sync and try to access an invalid index of this.entries. When
the entry cannot be found, we emit end and close, which can drop files
from reading.

(At least partially) fixes npm/npm#5082.

Credit: @evanlucas
Reviewed-By: @othiym23
PR-URL: https://github.com/npm/fstream/pull/50